Update initializiation of DynamoDB client
Update DynamoDBDocumentClient in core project to be a function
`initDynamoDbClient`.

The motivation for this is that lambda functions
created with SST via `sst.config.ts` at the root directory does fails to
find environment variables that are attempted to be configured
in the core project.

As a result, update the initializeTable function to receive the
dynamoDbClient as a prop. Also, revert the change to the core
project's build script