Remove baseApiGatewayHandler
Remove baseApiGatewayHandler due to issues in the lambda functions that
cause the thrown errors to not be displayed correctly in the logs. This
will be revisited at a later time.

As a result of the removal of baseApiGatewayHandler, refactor existing
endpoints for getNotes, getNote, and createNote so that these functions
will reuse common logic via `handleSuccessfulResponse` and
`handleResponseError`