Migrate from pnpm to yarn
Migrate from pnpm to yarn due to deployment issues with the current
stack where we use SST and Next.js for our frontend application.
The issue occurs after we deploy via `sst deploy` and the siteUrl
being accessed just shown an `Internal server error`. The exact
cause is not determined at this point in time.