Update createNote API to return created note
Update createNote API to return created note so that we can return this
as part of the response. The created note data can then be used for
optimistic updating on the client side. When the client knows that a
note is successfully persisted in the DB, it will update the cached note
that was created via optimistic updating so that the client will know
the noteId persisted in the backend.