Optimistic update note updates
Optimistic update note updates and consider scenario where
a user updates a note that still has a temporaryId, which is a note
that is optimistically updated via note creation.

Please note that although we expect the handling for updating
notes with temporary IDs to be create, there is a bug with this
where the final updated data is not as expected where changes
to a note between invocation of create endpoint and create success
fail to be retained in the cache.